packages/api/pyproject.toml
[project]
name = "lex-api"
version = "2.0.0"
description = "Lex the Computer — Product API (FastAPI)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30",
    "alembic>=1.14",
    "pydantic-settings>=2.7",
    "python-multipart>=0.0.18",
    "websockets>=14.0",
    "httpx>=0.28",
    "redis>=5.2",
    "honcho-ai>=2.1",
    "passlib[bcrypt]>=1.7",
    "python-jose[cryptography]>=3.3",
    "duckdb>=1.2",
    "stripe>=11.0",
    "docker>=7.1",
    "aiofiles>=24.1",
    "psutil>=6.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py312"
line-length = 100
